Big reorganization
Separate functionality to separate files
