+++ lib/workflow.rb
@@ -1,131 +1,13 @@
 require 'rubygems'
-require 'active_support/inflector'
+
+require 'workflow/specification'
+require 'workflow/adapters/active_record'
+require 'workflow/adapters/remodel'
+require 'workflow/draw'
 
 # See also README.markdown for documentation
 module Workflow
-
-  class Specification
-
-    attr_accessor :states, :initial_state, :meta,
-      :on_transition_proc, :before_transition_proc, :after_transition_proc, :on_error_proc
-
-    def initialize(meta = {}, &specification)
-      @states = Hash.new
-      @meta = meta
-      instance_eval(&specification)
-    end
-
-    def state_names
-      states.keys
-    end
-
-    private
-
-    def state(name, meta = {:meta => {}}, &events_and_etc)
-      # meta[:meta] to keep the API consistent..., gah
-      new_state = Workflow::State.new(name, self, meta[:meta])
-      @initial_state = new_state if @states.empty?
-      @states[name.to_sym] = new_state
-      @scoped_state = new_state
-      instance_eval(&events_and_etc) if events_and_etc
-    end
-
-    def event(name, args = {}, &action)
-      target = args[:transitions_to] || args[:transition_to]
-      raise WorkflowDefinitionError.new(
-        "missing ':transitions_to' in workflow event definition for '#{name}'") \
-        if target.nil?
-      @scoped_state.events[name.to_sym] =
-        Workflow::Event.new(name, target, (args[:meta] or {}), &action)
-    end
-
-    def on_entry(&proc)
-      @scoped_state.on_entry = proc
-    end
-
-    def on_exit(&proc)
-      @scoped_state.on_exit = proc
-    end
-
-    def after_transition(&proc)
-      @after_transition_proc = proc
-    end
-
-    def before_transition(&proc)
-      @before_transition_proc = proc
-    end
-
-    def on_transition(&proc)
-      @on_transition_proc = proc
-    end
-
-    def on_error(&proc)
-      @on_error_proc = proc
-    end
-  end
-
-  class TransitionHalted < Exception
-
-    attr_reader :halted_because
-
-    def initialize(msg = nil)
-      @halted_because = msg
-      super msg
-    end
-
-  end
-
-  class NoTransitionAllowed < Exception; end
-
-  class WorkflowError < Exception; end
-
-  class WorkflowDefinitionError < Exception; end
-
-  class State
-
-    attr_accessor :name, :events, :meta, :on_entry, :on_exit
-    attr_reader :spec
-
-    def initialize(name, spec, meta = {})
-      @name, @spec, @events, @meta = name, spec, Hash.new, meta
-    end
-
-    unless RUBY_VERSION < '1.9'
-      include Comparable
-
-      def <=>(other_state)
-        states = spec.states.keys
-        raise ArgumentError, "state `#{other_state}' does not exist" unless other_state.in? states
-        if states.index(self.to_sym) < states.index(other_state.to_sym)
-          -1
-        elsif states.index(self.to_sym) > states.index(other_state.to_sym)
-          1
-        else
-          0
-        end
-      end
-    end
-
-    def to_s
-      "#{name}"
-    end
-
-    def to_sym
-      name.to_sym
-    end
-  end
-
-  class Event
-
-    attr_accessor :name, :transitions_to, :meta, :action
-
-    def initialize(name, transitions_to, meta = {}, &action)
-      @name, @transitions_to, @meta, @action = name, transitions_to.to_sym, meta, action
-    end
-
-  end
-
-  module WorkflowClassMethods
+  module ClassMethods
     attr_reader :workflow_spec
 
     def workflow_column(column_name=nil)
@@ -164,7 +46,7 @@ module Workflow
     end
   end
 
-  module WorkflowInstanceMethods
+  module InstanceMethods
 
     def current_state
       loaded_state = load_workflow_state
@@ -336,118 +218,19 @@ module Workflow
     end
   end
 
-  module ActiveRecordInstanceMethods
-    def load_workflow_state
-      read_attribute(self.class.workflow_column)
-    end
-
-    # On transition the new workflow state is immediately saved in the
-    # database.
-    def persist_workflow_state(new_value)
-      if self.respond_to? :update_column
-        # Rails 3.1 or newer
-        update_column self.class.workflow_column, new_value
-      else
-        # older Rails; beware of side effect: other (pending) attribute changes will be persisted too
-        update_attribute self.class.workflow_column, new_value
-      end
-    end
-
-    private
-
-    # Motivation: even if NULL is stored in the workflow_state database column,
-    # the current_state is correctly recognized in the Ruby code. The problem
-    # arises when you want to SELECT records filtering by the value of initial
-    # state. That's why it is important to save the string with the name of the
-    # initial state in all the new records.
-    def write_initial_state
-      write_attribute self.class.workflow_column, current_state.to_s
-    end
-  end
-
-  module RemodelInstanceMethods
-    def load_workflow_state
-      send(self.class.workflow_column)
-    end
-
-    def persist_workflow_state(new_value)
-      update(self.class.workflow_column => new_value)
-    end
-  end
-
   def self.included(klass)
-    klass.send :include, WorkflowInstanceMethods
-    klass.extend WorkflowClassMethods
+    klass.send :include, InstanceMethods
+    klass.extend ClassMethods
+
     if Object.const_defined?(:ActiveRecord)
       if klass < ActiveRecord::Base
-        klass.send :include, ActiveRecordInstanceMethods
+        klass.send :include, Adapter::ActiveRecord::InstanceMethods
         klass.before_validation :write_initial_state
       end
     elsif Object.const_defined?(:Remodel)
-      if klass < Remodel::Entity
-        klass.send :include, RemodelInstanceMethods
+      if klass < Adapter::Remodel::Entity
+        klass.send :include, Remodel::InstanceMethods
       end
     end
-  end
-
-  # Generates a `dot` graph of the workflow.
-  # Prerequisite: the `dot` binary. (Download from http://www.graphviz.org/)
-  # You can use this method in your own Rakefile like this:
-  #
-  #     namespace :doc do
-  #       desc "Generate a graph of the workflow."
-  #       task :workflow => :environment do # needs access to the Rails environment
-  #         Workflow::create_workflow_diagram(Order)
-  #       end
-  #     end
-  #
-  # You can influence the placement of nodes by specifying
-  # additional meta information in your states and transition descriptions.
-  # You can assign higher `doc_weight` value to the typical transitions
-  # in your workflow. All other states and transitions will be arranged
-  # around that main line. See also `weight` in the graphviz documentation.
-  # Example:
-  #
-  #     state :new do
-  #       event :approve, :transitions_to => :approved, :meta => {:doc_weight => 8}
-  #     end
-  #
-  #
-  # @param klass A class with the Workflow mixin, for which you wish the graphical workflow representation
-  # @param [String] target_dir Directory, where to save the dot and the pdf files
-  # @param [String] graph_options You can change graph orientation, size etc. See graphviz documentation
-  def self.create_workflow_diagram(klass, target_dir='.', graph_options='rankdir="LR", size="7,11.6", ratio="fill"')
-    workflow_name = "#{klass.name.tableize}_workflow".gsub('/', '_')
-    fname = File.join(target_dir, "generated_#{workflow_name}")
-    File.open("#{fname}.dot", 'w') do |file|
-      file.puts %Q|
-digraph #{workflow_name} {
-  graph [#{graph_options}];
-  node [shape=box];
-  edge [len=1];
-      |
-
-      klass.workflow_spec.states.each do |state_name, state|
-        file.puts %Q{  #{state.name} [label="#{state.name}"];}
-        state.events.each do |event_name, event|
-          meta_info = event.meta
-          if meta_info[:doc_weight]
-            weight_prop = ", weight=#{meta_info[:doc_weight]}"
-          else
-            weight_prop = ''
-          end
-          file.puts %Q{  #{state.name} -> #{event.transitions_to} [label="#{event_name.to_s.humanize}" #{weight_prop}];}
-        end
-      end
-      file.puts "}"
-      file.puts
-    end
-    `dot -Tpdf -o'#{fname}.pdf' '#{fname}.dot'`
-    puts "
-Please run the following to open the generated file:
-
-open '#{fname}.pdf'
-
-"
   end
 end
