+++ test/main_test.rb
@@ -438,8 +438,8 @@ class MainTest < ActiveRecordTestCase
   test 'diagram generation' do
     begin
       $stdout = StringIO.new('', 'w')
-      Workflow::create_workflow_diagram(Order, '/tmp')
-      assert_match(/open.+\.pdf/, $stdout.string,
+      Workflow::Draw::workflow_diagram(Order, '/tmp')
+      assert_match(/run the following/, $stdout.string,
         'PDF should be generate and a hint be given to the user.')
     ensure
       $stdout = STDOUT
@@ -519,7 +519,7 @@ class MainTest < ActiveRecordTestCase
   test 'workflow graph generation' do
     Dir.chdir('/tmp') do
       capture_streams do
-        Workflow::create_workflow_diagram(Order, '/tmp')
+        Workflow::Draw::workflow_diagram(Order, '/tmp')
       end
     end
   end
@@ -527,7 +527,7 @@ class MainTest < ActiveRecordTestCase
   test 'workflow graph generation in a path with spaces' do
     `mkdir -p '/tmp/Workflow test'`
     capture_streams do
-      Workflow::create_workflow_diagram(Order,  '/tmp/Workflow test')
+      Workflow::Draw::workflow_diagram(Order, '/tmp/Workflow test')
     end
   end
