--- workflow.rb
@@ -1,1 +1,0 @@
-Dir["#{File.dirname(__FILE__)}/lib/*.rb"].each { |rb| require rb }
